# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DiscoveryShop Marketplace API",
    "version": "1.0.0",
    "description": "Professional API documentation for DiscoveryShop - Full-stack e-commerce marketplace platform with AI-powered discovery, dual-role accounts, real-time chat, and comprehensive admin controls.",
    "contact": {
      "name": "API Support",
      "email": "support@discoveryshop.local"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    },
    {
      "url": "https://api.discoveryshop.com",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List user orders",
        "description": "Get all orders placed by current user with pagination and filtering.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {"type": "string"},
            "description": "Filter by status"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer"},
            "description": "Page number"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create order",
        "description": "Create new order from current shopping cart. Triggers payment processing.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "shipping_address": {
                    "type": "object",
                    "properties": {
                      "street": {"type": "string"},
                      "city": {"type": "string"},
                      "state": {"type": "string"},
                      "zip": {"type": "string"},
                      "country": {"type": "string"}
                    }
                  },
                  "payment_method": {"type": "string"}
                },
                "required": ["shipping_address"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order details",
        "description": "Retrieve detailed order information with items, status, and tracking.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Order details retrieved"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/cancel": {
      "post": {
        "tags": ["Orders"],
        "summary": "Cancel order",
        "description": "Cancel pending order and refund to user. Can only cancel orders not yet shipped.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Order cancelled successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
...
  },
...
}